backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pymongo==4.6.0
langchain==0.1.0
langchain-core==0.1.0
langchain-community==0.0.10
langchain-mongodb==0.1.0
langchain-huggingface==0.0.1
groq==0.4.1
pdf2image==1.16.3
PyPDF2==3.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6